Delete on directory watch configurable

+++ src/main/scala/se/nimsa/sbx/directory/DirectoryWatchActor.scala
@@ -40,6 +40,7 @@ import scala.util.control.NonFatal
 
 class DirectoryWatchActor(watchedDirectory: WatchedDirectory,
                           storage: StorageService,
+                          deleteWatchedDirectory: Boolean,
                           metaDataServicePath: String = "../../MetaDataService",
                           anonymizationServicePath: String = "../../AnonymizationService")
                          (implicit val materializer: Materializer, timeout: Timeout) extends Actor with DicomStreamOps {
@@ -72,7 +73,8 @@ class DirectoryWatchActor(watchedDirectory: WatchedDirectory,
       .mapAsync(5) { path => // do import, delete if successful (might be permission issues here)
         storeDicomData(FileIO.fromPath(path), sbxSource, storage, Contexts.imageDataContexts, reverseAnonymization = true).map { metaData =>
         system.eventStream.publish(ImageAdded(metaData.image.id, sbxSource, !metaData.imageAdded))
-          Files.deleteIfExists(path)
+          if (deleteWatchedDirectory)
+            Files.deleteIfExists(path)
         }.recover {
           case NonFatal(e) =>
             SbxLog.error("Directory", s"Could not add file ${Paths.get(watchedDirectory.path).relativize(path)}: ${e.getMessage}")
@@ -103,5 +105,5 @@ class DirectoryWatchActor(watchedDirectory: WatchedDirectory,
 }
 
 object DirectoryWatchActor {
-  def props(watchedDirectory: WatchedDirectory, storage: StorageService)(implicit materializer: Materializer, timeout: Timeout): Props = Props(new DirectoryWatchActor(watchedDirectory, storage))
+  def props(watchedDirectory: WatchedDirectory, storage: StorageService, deleteWatchedDirectory: Boolean)(implicit materializer: Materializer, timeout: Timeout): Props = Props(new DirectoryWatchActor(watchedDirectory, storage, deleteWatchedDirectory))
 }
